% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{readIt}
\alias{readIt}
\title{function that grabs the data from the website and converts to R data frame}
\usage{
readIt(address)
}
\arguments{
\item{address}{address of the data in json format}
}
\value{
data in R-data.frame
}
\description{
function that grabs the data from the website and converts to R data frame
}

